- Can remove question from questionnaire - A test in GetModulesStepdefs stopped working

+++ src/main/java/fr/uca/springbootstrap/controllers/QuestionnaireController.java
@@ -1,6 +1,8 @@
 package fr.uca.springbootstrap.controllers;
 
+import fr.uca.springbootstrap.models.Module;
 import fr.uca.springbootstrap.models.Questionnaire;
+import fr.uca.springbootstrap.models.User;
 import fr.uca.springbootstrap.models.questions.QCM;
 import fr.uca.springbootstrap.models.questions.Question;
 import fr.uca.springbootstrap.payload.request.QuestionRequest;
@@ -15,6 +17,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.security.Principal;
+import java.util.Optional;
 
 @CrossOrigin(origins = "*", maxAge = 3600)
 @RestController
@@ -79,9 +82,9 @@ public class QuestionnaireController {
 //    }
 
 
-    @PostMapping("/{questionnaire_id}/qcm")
-    @PreAuthorize("hasRole('TEACHER') or hasRole('ADMIN')")
-    public ResponseEntity<?> addQcm(Principal principal,
+    @PostMapping("/{questionnaire_id}/question")
+    @PreAuthorize("hasRole('TEACHER')")
+    public ResponseEntity<?> addQuestion(Principal principal,
                                          @Valid @RequestBody QuestionRequest questionRequest,
                                          @PathVariable("module_id") long module_id,
                                          @PathVariable("questionnaire_id") long questionnaire_id) {
@@ -94,20 +97,22 @@ public class QuestionnaireController {
                     .body(new MessageResponse("Error: User does not exist."));
         }
         Questionnaire questionnaire = questionnaireRepository.findById(questionnaire_id).get();
-        Question question = new QCM(questionRequest.getNumber(), questionRequest.getName(), questionRequest.getDescription(), questionnaire);
+        Question question = new QCM(questionRequest.getNumber(), questionRequest.getName(), questionRequest.getDescription());
+
+        questionRepository.save(question);
         questionnaire.addQuestion(question);
         questionnaireRepository.save(questionnaire);
-        questionRepository.save(question);
 
         return ResponseEntity.ok(new MessageResponse("Question successfully added."));
     }
 
-    @DeleteMapping("/{questionnaire_id}/qcm/{question_id}")
-    @PreAuthorize("hasRole('TEACHER') or hasRole('ADMIN')")
-    public ResponseEntity<?> removeQcm(Principal principal,
+    @DeleteMapping("/{questionnaire_id}/question/{question_id}")
+    @PreAuthorize("hasRole('TEACHER')")
+    public ResponseEntity<?> removeQuestion(Principal principal,
                                             @PathVariable("module_id") long module_id,
                                             @PathVariable("questionnaire_id") long questionnaire_id,
                                             @PathVariable("question_id") long question_id) {
+
         if (!userRepository.existsByUsername(principal.getName())) {
             return ResponseEntity.badRequest()
                     .body(new MessageResponse("Error: User does not exist."));
@@ -116,9 +121,21 @@ public class QuestionnaireController {
             return ResponseEntity.badRequest()
                     .body(new MessageResponse("Error: Question does not exist."));
         }
+        else if (!moduleRepository.existsById(module_id)) {
+            return ResponseEntity.badRequest()
+                    .body(new MessageResponse("Error: Module does not exist."));
+        }
+
+        Module module = moduleRepository.findById(module_id).get();
+        User user = userRepository.findByUsername(principal.getName()).get();
+        if (!module.containsParticipant(user)) {
+            return ResponseEntity.badRequest()
+                    .body(new MessageResponse("Error: user does not belong to this module."));
+        }
 
         Question question = questionRepository.findById(question_id).get();
         Questionnaire questionnaire = questionnaireRepository.findById(questionnaire_id).get();
+
         questionnaire.removeQuestion(question);
         questionRepository.delete(question);
         questionnaireRepository.save(questionnaire);
